refactor - move clientMemberMutation to hooks
멤버리스트 내의 clientMemberMutation 함수를 useClientQueries 파일로 이동
