+++ src/hooks/useClientsQueries.ts
@@ -1,5 +1,6 @@
-import { useQuery } from "@tanstack/react-query";
-import { getClientByProject } from "src/api/User";
+import { useMutation, useQuery } from "@tanstack/react-query";
+import { queryClient } from "src/App";
+import { getClientByProject, updateUser } from "src/api/User";
 
 const useClientsQueries = (clientId: string) => {
   const { data: client } = useQuery(
@@ -13,8 +14,19 @@ const useClientsQueries = (clientId: string) => {
     }
   );
 
+  const clientMembersMutation = useMutation(
+    ({ updatedData, userId }: { updatedData: object; userId: string }) =>
+      updateUser({ updatedData, userId }),
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(["clients"]);
+      },
+    }
+  );
+
   return {
     client,
+    clientMembersMutation,
   };
 };
